chore(ci): only run snyk, fossa and semgrep on pushes to main
This is because external contributors do not have access to the required
secrets necessary for running these workflows, causing them to fail.

This causes external contributors confusion where they believe the failure
was triggered by a change they made.

Instead for FOSSA and snyk we will rely on the github integration which
does not need access to defined secrets, though does detect less issues.

These workflows are kept in main so that we are able to detect any issues
missed by the integration,