fix: read env vars for migrate command
This is the opposite of DRY. However, I think it is okay here. Since version does not belong in the app config, I don't think the migrate command should be reading values from the app `config.yaml`. Hence, I didn't refactor the flags from run to also apply here.

Also OPENFGA_VERSION is a weird name for the version env var. But it is the appropriate name for the flag. So... :shrug: