feat(migrate): Add env vars and username/password overwrite for migrate tool
Before, openfga run had the possibility to overwrite the username / password present
in the run command, but not in the migrate command.

This commit fixes it by adding the flags and the corresponding behavior to
the migrate command. Furthermore it adds the env variable handling from
run also to migrate