Publish OpenFGA to GHCR as well
This proposes changes to the goreleaser configuration and the
release.yaml GitHub workflow to allow the process to release both to
DockerHub (as before) and to GHCR. The name of the package will be
`ghcr.io/openfga/openfga` in the registry.

Note that an OpenFGA maintainer with admin permissions will need to make
the package public explicitly, since it will be created with "private"
visibility by default.

The need for this comes from the fact that DockerHub has very aggressive
rate limits, which can be problematic for environments that use
ephemeral node instances (SPOT instances) which tend to pull more often.

Fortunately, GHCR is free for open source projects and has more lenient
and reasonable rate limits.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>