feat(server): add gRPC header for query duration in server methods
This commit adds a gRPC header (`fga-query-duration-ms`) to track the duration of various server methods. The header is set with the time taken to execute the request in milliseconds. This change affects the following methods:

- `ListObjects`
- `StreamedListObjects`
- `Read`
- `Write`
- `Check`
- `Expand`
- `ReadAuthorizationModel`
- `WriteAuthorizationModel`
- `ReadAuthorizationModels`
- `WriteAssertions`
- `ReadAssertions`
- `ReadChanges`
- `CreateStore`
- `DeleteStore`
- `GetStore`
- `ListStores`

The duration is calculated at the start and end of each method, and the header is sent using `grpc.SendHeader`.