ci: add action to enforce adding changelog entries
This will run on all PRs and report a failure if they do not modify the changelog, if we do not
need a changelog entry for that specific PR we can add a Skip-Changelog label on the PR to skip
the check. We can customise the label if we don't like that one.