fix db connections from prematurely closing
when the context that was passed to QueryContext is canceled,
the resulting *Rows is also canceled and cannot be added back
to the pool because it is in an unusable state.

this commit removes the cancel from the context, giving *Rows
time to complete or be closed, and the connection reused.