remove closed parameter from the shared iterator.
a stopped shared iterator should never be cloned.
by enforcing that cloning can only be done from a live shared iterator,
we ensure that the reference counter for that shared iterator is greater
than 0 at the time of cloning; meaning that the inner iterator has not
yet had Stop called on it.