add necessary valid context to tests
in algorithms that evaluate all possible model paths, this context
is necessary to be present. if, during evaluation, a tuple is
encountered which specifies a condition, but both the tuple and the
request are missing the appropriate context parameters, the result will
be an error.

the only reason that the current algorithm does not fail on these tests
is because it happens to not encounter such a tuple due to using the
check algorithm in its implementation.